+++ src/daemon/influxd.go
@@ -84,7 +84,7 @@ func setupLogging(loggingLevel string) {
 }
 
 func main() {
-	fileName := flag.String("config", "config.json.sample", "Config file")
+	fileName := flag.String("config", "config.toml.sample", "Config file")
 	wantsVersion := flag.Bool("v", false, "Get version number")
 	resetRootPassword := flag.Bool("reset-root", false, "Reset root password")
 	pidFile := flag.String("pidfile", "", "the pid file")